# docs/api_doc/swagger.json
{"openapi":"3.1.0","info":{"title":"'deviaje-users-and-auth'","description":"'deviaje-users-and-auth'","contact":{"name":"John Doe","email":"dumy@dumy"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:9060","description":"'deviaje-users-and-auth'"}],"paths":{"/api/users/{id}":{"get":{"tags":["user-controller"],"operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}},"put":{"tags":["user-controller"],"operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserPut"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}},"delete":{"tags":["user-controller"],"operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}}}}},"/api/users":{"get":{"tags":["user-controller"],"operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}}}}}}},"post":{"tags":["user-controller"],"operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/api/users/{id}/activate":{"post":{"tags":["user-controller"],"operationId":"activateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}}}}},"/api/auth/signup":{"post":{"tags":["auth-controller"],"operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}}}}},"/api/auth/reset-password":{"post":{"tags":["auth-controller"],"operationId":"resetPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}}}}},"/api/auth/refresh-token":{"post":{"tags":["auth-controller"],"operationId":"refreshToken","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshTokenRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JwtResponse"}}}}}}},"/api/auth/logout":{"post":{"tags":["auth-controller"],"operationId":"logoutUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"authenticateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/JwtResponse"}}}}}}},"/api/auth/forgot-password":{"post":{"tags":["auth-controller"],"operationId":"forgotPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgotPasswordRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}}}}},"/api/auth/change-password":{"post":{"tags":["auth-controller"],"operationId":"changePassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordChangeRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}}}}},"/api/users/role/{role}":{"get":{"tags":["user-controller"],"operationId":"getUsersByRole","parameters":[{"name":"role","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}}}}}}}}},"components":{"schemas":{"UserPut":{"properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"phone":{"type":"string"},"birthDate":{"type":"string","format":"date"},"dni":{"type":"string"},"dniTypeId":{"type":"integer","format":"int32"},"roleIds":{"type":"array","items":{"type":"integer","format":"int32"},"uniqueItems":true},"membershipId":{"type":"integer","format":"int32"},"avatarUrl":{"type":"string"},"active":{"type":"boolean"},"lastUpdatedUser":{"type":"integer","format":"int32"}}},"MembershipDto":{"properties":{"id":{"type":"integer","format":"int32"},"description":{"type":"string"},"cost":{"type":"number"},"discountPercentage":{"type":"number"}}},"UserDto":{"properties":{"id":{"type":"integer","format":"int32"},"username":{"type":"string"},"email":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"phone":{"type":"string"},"birthDate":{"type":"string","format":"date"},"dni":{"type":"string"},"dniType":{"type":"string"},"active":{"type":"boolean"},"avatarUrl":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}},"membership":{"$ref":"#/components/schemas/UserMembershipDto"}}},"UserMembershipDto":{"properties":{"id":{"type":"integer","format":"int32"},"userId":{"type":"integer","format":"int32"},"membership":{"$ref":"#/components/schemas/MembershipDto"},"currentPoints":{"type":"integer","format":"int32"},"startDate":{"type":"string","format":"date-time"},"endDate":{"type":"string","format":"date-time"}}},"UserCreateRequest":{"properties":{"username":{"type":"string","maxLength":50,"minLength":3},"email":{"type":"string","minLength":1},"password":{"type":"string","maxLength":40,"minLength":6},"firstName":{"type":"string"},"lastName":{"type":"string"},"phone":{"type":"string"},"birthDate":{"type":"string","format":"date"},"dni":{"type":"string"},"dniTypeId":{"type":"integer","format":"int32"},"roleIds":{"type":"array","items":{"type":"integer","format":"int32"},"uniqueItems":true},"createdUser":{"type":"integer","format":"int32"}}},"MessageResponse":{"properties":{"message":{"type":"string"}}},"SignupRequest":{"properties":{"username":{"type":"string","maxLength":50,"minLength":3},"email":{"type":"string","minLength":1},"password":{"type":"string","maxLength":40,"minLength":6},"firstName":{"type":"string"},"lastName":{"type":"string"},"phone":{"type":"string"},"birthDate":{"type":"string","format":"date"},"dni":{"type":"string"},"dniTypeId":{"type":"integer","format":"int32"}}},"ResetPasswordRequest":{"properties":{"token":{"type":"string","minLength":1},"newPassword":{"type":"string","maxLength":40,"minLength":6},"confirmPassword":{"type":"string","minLength":1}}},"RefreshTokenRequest":{"properties":{"refreshToken":{"type":"string","minLength":1}}},"JwtResponse":{"properties":{"token":{"type":"string"},"refreshToken":{"type":"string"},"type":{"type":"string"},"id":{"type":"integer","format":"int32"},"username":{"type":"string"},"email":{"type":"string"},"roles":{"type":"array","items":{"type":"string"}}}},"LoginRequest":{"properties":{"username":{"type":"string","minLength":1},"password":{"type":"string","minLength":1}}},"ForgotPasswordRequest":{"properties":{"email":{"type":"string","minLength":1}}},"PasswordChangeRequest":{"properties":{"currentPassword":{"type":"string","minLength":1},"newPassword":{"type":"string","maxLength":2147483647,"minLength":8},"confirmPassword":{"type":"string","minLength":1}}}}}}